requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2022.2.1
sqlparse~=0.4.2

pip~=22.0.2
wheel~=0.37.1
cryptography~=3.4.8
Pillow~=9.0.1
Jinja2~=3.1.2
keyring~=23.5.0
colorama~=0.4.4
PyYAML~=5.4.1
requests~=2.25.1
environ~=1.0
setuptools~=59.6.0
future~=0.18.2
django-debug-toolbar~=3.6.0
coreapi~=2.3.3
uritemplate~=4.1.1
coreschema~=0.0.4
django-filter~=22.1